stock_processor/requirements.txt
grpcio
grpcio-tools
kafka-python
pymongo